requirements.txt
gradio
interpret
pandas
numpy
scikit-learn
matplotlib
Pillow
